test: refactor WTR config files
