+++ web-test-runner.config.ci-firefox.js
@@ -9,7 +9,7 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
-import { firefox } from './web-test-runner.config.js';
+import { firefox } from './web-test-runner.utils.js';
 import standard from './web-test-runner.config.ci.js';
 
 standard.browsers = [firefox];
